requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
python-dotenv==1.1.1
setuptools==80.9.0
sqlparse==0.5.3
wheel==0.45.1
